+++ matflow/validation.py
@@ -5,68 +5,173 @@ import warnings
 from matflow.errors import UnsatisfiedSchemaError
 
 
-def validate_function_args(func, func_type, expected_args=None, num_args=None):
-    """
+def validate_input_mapper_func(func, task_inputs):
+    """Using `inspect`, validate an input mapper callable from a Matflow extension.
+
     Parameters
     ----------
     func : callable
-    func_type : str
-        One of "input_mapper", "output_mapper", ...
-    expected_args : list, optional
-        Must be specified if `func_type` is "input_mapper" or "func_mapper".
-    num_args : int, optional
-        Must be specified if `func_type` is "output_mapper".
+    task_inputs : list of str
+        List of the input name aliases associated with the task schema.
+
+    Notes
+    -----
+    Checks performed on `func`:
+      - check the first argument is named "path"; raise `TypeError` if not;
+      - check for one or more additional arguments which are named according to
+        a subset of task parameters (passed in `task_inputs`).
 
     """
 
-    func_type_fmt = func_type.replace('_', ' ')
     func_params = inspect.signature(func).parameters
 
-    if func_type == 'input_mapper':
-        # Check first argument must be "path":
-        first_arg = list(func_params.items())[0]
-        if first_arg[0] != 'path':
-            msg = (f'The first parameter of an {func_type_fmt} function must be "path" '
-                   f'but for {func.__name__} is actually "{first_arg[0]}".')
-            raise TypeError(msg)
-        else:
-            func_params = dict(func_params)
-            del func_params[first_arg[0]]
+    # Check first argument must be "path":
+    first_arg = list(func_params.items())[0]
+    if first_arg[0] != 'path':
+        msg = (f'The first parameter of an input mapper function must be "path" '
+               f'but for {func.__name__} is actually "{first_arg[0]}".')
+        raise TypeError(msg)
+    else:
+        # Remove "path" from argument list, for further analysis:
+        func_params = dict(func_params)
+        del func_params[first_arg[0]]
 
-    elif func_type == 'output_mapper':
-        # Check num args:
-        if len(func_params) != num_args:
-            msg = (f'There are {len(func_params)} parameters in output mapping function '
-                   f'"{func.__name__}", but {num_args} are expected.')
-            warnings.warn(msg)
+    bad_params = list(set(func_params) - set(task_inputs))
+    if bad_params:
+        bad_params_fmt = ', '.join([f'"{i}"' for i in bad_params])
+        msg = (f'The following schema input parameters are not compatible with the '
+               f'input mapper function "{func.__name__}": {bad_params_fmt}.')
+        raise TypeError(msg)
 
-    if func_type in ['input_mapper', 'func_mapper']:
-        non_defaulted = [k for k, v in func_params.items() if v.default == inspect._empty]
-        # Parameters for which their are no defaults must be supplied by schema:
-        miss_params = list(set(non_defaulted) - set(expected_args))
-        if miss_params:
-            miss_params_fmt = ', '.join([f'"{i}"' for i in miss_params])
-            msg = (f'The following parameters in {func_type_fmt} function '
-                   f'"{func.__name__}" have no default value and are not specified in '
-                   f'the schema: {miss_params_fmt}.')
-            raise TypeError(msg)
 
-        # All parameters in the schema must appear in the function:
-        bad_params = list(set(expected_args) - set(func_params))
-        if bad_params:
-            bad_params_fmt = ', '.join([f'"{i}"' for i in bad_params])
-            msg = (f'The following schema input parameters are not compatible with the '
-                   f'{func_type_fmt} function "{func.__name__}": {bad_params_fmt}.')
-            raise TypeError(msg)
+def validate_output_mapper_func(func, num_file_paths, option_names):
+    """Using `inspect`, validate an output mapper callable from a Matflow extension.
+
+    Parameters
+    ----------
+    func : callable
+    num_file_paths : int
+        Number of output files specified in the schema's output map.
+    option_names : list of str
+        List of the names of output map options.
+
+    Notes
+    -----
+    Checks performed on `func`:
+      - After the first `num_file_paths` arguments, check the remaining arguments names
+        coincide exactly with `option_names`.
+
+    """
+
+    func_params = inspect.signature(func).parameters
+
+    # Check num args first
+    exp_num_params = num_file_paths + len(option_names)
+    if len(func_params) != exp_num_params:
+        msg = (
+            f'The output mapper function "{func.__name__}" does not have the expected '
+            f'number of arguments: found {len(func_params)} but expected '
+            f'{exp_num_params} ({num_file_paths} file path(s) + {len(option_names)} '
+            f'options parameters).'
+        )
+        raise TypeError(msg)
+
+    # Check option names:
+    opt_params = list(func_params.items())[num_file_paths:]
+    opt_params_func = [i[0] for i in opt_params]
+
+    miss_params = list(set(option_names) - set(opt_params_func))
+    bad_params = list(set(opt_params_func) - set(option_names))
+
+    if bad_params:
+        bad_params_fmt = ', '.join([f'"{i}"' for i in bad_params])
+        msg = (f'The following arguments in the output mapper function "{func.__name__}" '
+               f'are not known to be output mapper options: {bad_params_fmt}.')
+        raise TypeError(msg)
+
+    if miss_params:
+        miss_params_fmt = ', '.join([f'"{i}"' for i in miss_params])
+        msg = (f'The following output mapper options are missing from the signature of '
+               f'the output mapper function "{func.__name__}": {miss_params_fmt}.')
+        raise TypeError(msg)
+
+
+def validate_func_mapper_func(func, task_inputs):
+    """Using `inspect`, validate an input mapper callable from a Matflow extension.
+
+    Parameters
+    ----------
+    func : callable
+    task_inputs : list of str
+        List of the input name aliases associated with the task schema.
+
+    Notes
+    -----
+    Checks performed on `func`:
+      - check function arguments are named according to all task parameters (passed in
+      `task_inputs`).
+
+    """
+
+    func_params = inspect.signature(func).parameters
+
+    bad_params = list(set(func_params) - set(task_inputs))
+    miss_params = list(set(task_inputs) - set(func_params))
+
+    if bad_params:
+        bad_params_fmt = ', '.join([f'"{i}"' for i in bad_params])
+        msg = (f'The function mapper function "{func.__name__}" contains the following '
+               f'arguments that are not consistent with the schema: {bad_params_fmt}.')
+        raise TypeError(msg)
+
+    if miss_params:
+        miss_params_fmt = ', '.join([f'"{i}"' for i in miss_params])
+        msg = (f'The following task inputs are missing from the signature of the '
+               f'function mapper function "{func.__name__}": {miss_params_fmt}.')
+        raise TypeError(msg)
 
 
 def validate_task_schemas(task_schemas, task_input_map, task_output_map, task_func_map):
+    """
+    Determine whether each task schema is valid.
+
+    Parameters
+    ----------
+    task_schemas : dict of (tuple : TaskSchema)
+        Dict keys are (task_name, task_method, software).
+    task_input_map : dict of (tuple : dict of (str : callable))
+        Outer dict keys are (task_name, task_method, software); inner dicts map a string
+        input file name to a MatFlow extension callable which writes that input file.
+    task_output_map : dict of (tuple : dict of (str : callable))
+        Outer dict keys are (task_name, task_method, software); inner dicts map a string
+        output name to a MatFlow extension callable which return that output.
+    task_func_map : dict of (tuple : callable)
+        Dict keys are (task_name, task_method, software); values are MatFlow extension
+        callables.
+
+    Returns
+    -------
+    schema_is_valid : dict of (tuple : tuple of (bool, str))
+        Dict keys are (task_name, task_method, software); values are tuples whose first
+        values are boolean values indicating if a given schema is valid. If False, this
+        indicates that one of extension functions (input map, output map or function map)
+        is missing. Note that this function does not raise any exception in this case ---
+        but the task schema will be noted as invalid. The second value of the dict value
+        tuple is a string description of the reason why the schema is invalid.
+
+    Raises
+    ------
+    UnsatisfiedSchemaError
+        Raised if any of the extension callables (input/output/func maps) are not
+        consistent with their associated task schema.
+
+    """
 
     schema_is_valid = {}
 
     for key, schema in task_schemas.items():
 
-        schema_is_valid.update({key: True})
+        schema_is_valid.update({key: (True, '')})
 
         key_msg = (f'Unresolved task schema for task "{schema.name}" with method '
                    f'"{schema.method}" and software "{schema.implementation}".')
@@ -80,20 +185,18 @@ def validate_task_schemas(task_schemas, task_input_map, task_output_map, task_fu
             )
 
             if not extension_inp_maps:
-                schema_is_valid.update({key: False})
+                reason = (f'No input map function found for input map that generates file'
+                          f' "{inp_map["file"]}". ')
+                schema_is_valid[key] = (False, schema_is_valid[key][1] + reason)
                 continue
             else:
                 inp_map_func = extension_inp_maps.get(inp_map['file'])
                 if not inp_map_func:
                     raise UnsatisfiedSchemaError(msg)
 
-            # Validate input map inputs against func args:
+            # Validate signature of input map function:
             try:
-                validate_function_args(
-                    func=inp_map_func,
-                    func_type='input_mapper',
-                    expected_args=inp_map['inputs'],
-                )
+                validate_input_mapper_func(inp_map_func, inp_map['inputs'])
             except TypeError as err:
                 raise UnsatisfiedSchemaError(key_msg + ' ' + str(err)) from None
 
@@ -106,19 +209,21 @@ def validate_task_schemas(task_schemas, task_input_map, task_output_map, task_fu
             )
 
             if not extension_out_maps:
-                schema_is_valid.update({key: False})
+                reason = (f'No output map function found for output map that generates '
+                          f'output "{out_map["output"]}". ')
+                schema_is_valid[key] = (False, schema_is_valid[key][1] + reason)
                 continue
             else:
                 out_map_func = extension_out_maps.get(out_map['output'])
                 if not out_map_func:
                     raise UnsatisfiedSchemaError(msg)
 
-            # Validate number of output map args
+            # Validate signature of output map function:
             try:
-                validate_function_args(
+                validate_output_mapper_func(
                     func=out_map_func,
-                    func_type='output_mapper',
-                    num_args=(len(out_map['files']) + len(out_map.get('options', {}))),
+                    num_file_paths=len(out_map['files']),
+                    option_names=[i['name'] for i in out_map.get('options', [])],
                 )
             except TypeError as err:
                 raise UnsatisfiedSchemaError(key_msg + ' ' + str(err)) from None
@@ -127,15 +232,13 @@ def validate_task_schemas(task_schemas, task_input_map, task_output_map, task_fu
 
             func = task_func_map.get(key)
             if not func:
-                schema_is_valid.update({key: False})
+                reason = 'No function mapper function found. '
+                schema_is_valid[key] = (False, schema_is_valid[key][1] + reason)
                 continue
 
+            # Validate signature of func mapper function:
             try:
-                validate_function_args(
-                    func=func,
-                    func_type='func_mapper',
-                    expected_args=schema.input_aliases,
-                )
+                validate_func_mapper_func(func, schema.input_aliases)
             except TypeError as err:
                 raise UnsatisfiedSchemaError(key_msg + ' ' + str(err)) from None